Cast pointer to uintptr_t instead of unsigned long. Because long could be 32bit on other platforms (ie. Windows) regardless of architecture.

+++ src/utils/gravity_json.c
@@ -433,7 +433,7 @@ static int new_value (json_state * state,
             values_size = sizeof (*value->u.object.values) * value->u.object.length;
 
             if (! (value->u.object.values = (json_object_entry *) json_alloc
-                  (state, values_size + ((unsigned long) value->u.object.values), 0)) )
+                  (state, values_size + ((uintptr_t) value->u.object.values), 0)) )
             {
                return 0;
             }
